% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{A1inv}
\alias{A1inv}
\title{Inverse of A1 function.
This is available in the circular package but we have recreated it here from
Bays' code}
\usage{
A1inv(r)
}
\description{
Inverse of A1 function.
This is available in the circular package but we have recreated it here from
Bays' code
}
