% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysing.R
\name{analyse_precision}
\alias{analyse_precision}
\title{Obtain measures of precision}
\usage{
analyse_precision(
  data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  set_size_var = NULL,
  condition_var = NULL
)
}
\arguments{
\item{data}{A data frame with columns containing: participant identifier
(declared via variable 'id_var'); the participants' response per trial ('response_var'); the
target value ('target_var'); and, if applicable, the set size of each
response ('set_size_var'), and the condition of each response
('condition_var').}

\item{unit}{The unit of measurement in the data frame: "degrees"
(measurement is in degrees, from 0 to 360); "degrees_180 (measurement is in
degrees, but limited to 0 to 180); or "radians" (measurement is in radians,
from pi to 2 * pi, but could also be already in -pi to pi).}

\item{id_var}{The quoted column name coding for participant id. If the data is from
a single participant (i.e., there is no id column) set to NULL.}

\item{response_var}{The quoted column name coding for the participants' responses}

\item{target_var}{The quoted column name coding for the target value.}

\item{set_size_var}{The quoted column name (if applicable) coding for the set
size of each response.}

\item{condition_var}{The quoted column name (if applicable) coding for the
condition of each response.}
}
\value{
\code{precision} The reciprocal of the circular standard deviation of
the response error, with correction for guessing.

\code{bias} The bias of the precision estimate.
}
\description{
Returns participant-level data for precision estimates ready for inferential
analysis. Note that the function does not actually conduct the analysis.
}
\details{
Precision is defined as the reciprocal of the circular standard deviation of
the response error corrected for guessing.
}
\examples{
# load an example data frame
data(bays2009_full)

# calculate the precision per condition per set size
precision_data <- analyse_precision(data = bays2009_full,
                                    unit = "radians",
                                    condition_var = "duration",
                                    set_size_var = "set_size")

}
