% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{berry_2019}
\alias{berry_2019}
\title{Data set from Berry et al. (2019)}
\format{
A data frame with 3600 rows and 6 variables:
\describe{
  \item{id}{participant identification}
  \item{condition}{condition of experiment: whether the task was completed
  under single-task or dual-task conditions}
  \item{target_ori}{the orientation of the target in degrees (1-180)}
  \item{response_ori}{the participant's recollection of the target
  orientation in degrees (1-180)}
  \item{non_target_1}{the orientation of the first non-target in degrees
  (1-180)}
  \item{non_target_2}{the orientation of the second non-target in degrees
  (1-180)}
}
}
\source{
The data set is publicly available on the Open Science Framework:
https://osf.io/59c4g/
}
\usage{
berry_2019
}
\description{
A data set including data from 30 participants in a continuous report
visual short-term memory experiment. The stimuli were oriented bars within
the range 1-180 degrees. The experiment had a set size of 3.
}
\references{
Berry. E.D.J., Allen, R.J., Waterman, A.H., & Logie, R.H. (2019). The
   effect of a verbal concurrent task on visual precision in working
   memory. Experimental Psychology, 66, (77-85).
}
\keyword{datasets}
