% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cmean}
\alias{cmean}
\title{Calculate the circular equivalent of the mean}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
cmean(x)
}
\description{
Calculate the circular equivalent of the mean
}
