% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{components_model_pdf_em}
\alias{components_model_pdf_em}
\title{Calculate the likelihood function of the components model fitting via
expectation maximisation.}
\usage{
components_model_pdf_em(response, target, non_targets, start_parms = NULL)
}
\description{
It is not expected that this function be called by the user.
}
