% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{components_model_pdf_gd}
\alias{components_model_pdf_gd}
\title{Calculate the likelihood function of the mixture model.}
\usage{
components_model_pdf_gd(
  response,
  target,
  non_targets,
  start_parms = NULL,
  min_parms,
  max_parms
)
}
\description{
It is not expected that this function be called by the user.
}
