% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cstd}
\alias{cstd}
\title{Calculate the circular equivalent of standard deviation}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
cstd(x)
}
\description{
Calculate the circular equivalent of standard deviation
}
