% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{degrees_to_radians}
\alias{degrees_to_radians}
\title{Transform degrees into radians}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
degrees_to_radians(deg)
}
\arguments{
\item{deg}{Degree value to transform into radians}
}
\description{
A function to transform degrees into radians
}
