% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_level}
\alias{fit_level}
\title{fit model to a single level}
\usage{
fit_level(
  data,
  components = 3,
  id_var = "id",
  response_var = "response",
  target_var = "target",
  non_target_var = non_target_var,
  set_size = 1
)
}
\description{
fit model to a single level
}
