% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_level_slots}
\alias{fit_level_slots}
\title{Fit slots model to a single level. This fits both the slots model and the
slots plus averaging model.}
\usage{
fit_level_slots(
  data,
  model,
  id_var,
  response_var,
  target_var,
  set_size_var,
  return_fit,
  fit_method
)
}
\description{
This wrapper function is called by the \code{fit_mixtur} function to fit the
models to a single level from the data frame. It is not expected that this
function be called by the user.
}
