% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_mixtur}
\alias{fit_mixtur}
\title{Fit the mixture model. This is the function to be called by the user
TODO: Documentation!}
\usage{
fit_mixtur(
  data,
  components = 3,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  non_target_var = "NULL",
  set_size_var = "NULL",
  condition_var = "NULL"
)
}
\description{
Fit the mixture model. This is the function to be called by the user
TODO: Documentation!
}
