% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_model}
\alias{fit_model}
\title{Fit the mixture model.}
\usage{
fit_model(
  response,
  target,
  non_targets = replicate(NROW(response), 0),
  return.ll = TRUE
)
}
\description{
This is the function that is called by the wrapper function
\code{fit_level}. It is not expected that this function be called by the
user.
}
