% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_slots_gd}
\alias{fit_slots_gd}
\title{fit the slots model via gradient descent.}
\usage{
fit_slots_gd(model, slot_data, return.ll = TRUE)
}
\description{
This is the function that is called by the wrapper function
\code{fit_level_slots}. It is not expected that this function be called by
the user.
}
