% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{fixed_angle_simulate_mixtur}
\alias{fixed_angle_simulate_mixtur}
\title{simulate the mixture model with fixed angle separation.
This is NOT the function to be called by the user}
\usage{
fixed_angle_simulate_mixtur(
  n_trials,
  K,
  p_t,
  p_n,
  p_u,
  set_size = 4,
  min_angle_distance = 20
)
}
\description{
simulate the mixture model with fixed angle separation.
This is NOT the function to be called by the user
}
