% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{get_component_response}
\alias{get_component_response}
\title{get the mixtur model's responses}
\usage{
get_component_response(trial_data, set_size, kappa, p_t, p_n, p_u)
}
\description{
get the mixtur model's responses
}
