% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_precision_single}
\alias{get_precision_single}
\title{Obtain the precision of a single condition}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
get_precision_single(error, target = 0)
}
\description{
Obtain the precision of a single condition
}
