% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{likelihood_function}
\alias{likelihood_function}
\title{Calculate the likelihood function of the mixture model.}
\usage{
likelihood_function(response, target, non_targets, start_parms = NULL)
}
\description{
It is not expected that this function be called by the user.
}
