% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logspace}
\alias{logspace}
\title{Obtain logarithmically spaced vectors
logspace function for logarithmically spaced vectors
soure: http://r.789695.n4.nabble.com/logarithmic-seq-tp900431p900433.html
#' @export}
\usage{
logspace(a, b, n)
}
\description{
Obtain logarithmically spaced vectors
logspace function for logarithmically spaced vectors
soure: http://r.789695.n4.nabble.com/logarithmic-seq-tp900431p900433.html
#' @export
}
