% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{mixtur_model_pdf}
\alias{mixtur_model_pdf}
\title{Calculate the likelihood function of the mixture model.}
\usage{
mixtur_model_pdf(response, target, non_targets, start_parms = NULL)
}
\description{
It is not expected that this function be called by the user.
}
