% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{mixtur_model_pdf_optim}
\alias{mixtur_model_pdf_optim}
\title{Calculate the likelihood function of the mixture model.}
\usage{
mixtur_model_pdf_optim(
  response,
  target,
  non_targets,
  start_parms = NULL,
  min_parms,
  max_parms
)
}
\description{
It is not expected that this function be called by the user.
}
