% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{model_comparison}
\alias{model_comparison}
\title{Conduct formal model comparison.}
\usage{
model_comparison(
  data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  non_target_var,
  set_size_var = NULL,
  condition_var = NULL
)
}
\description{
This function will fit both the 2- and 3-component models to the data
and provide an assessment of which fits best.
}
