% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_error_non_targets}
\alias{plot_error_non_targets}
\title{Plot response error of behavioural target relative to non-target values
#'
Function to plot the response error in behavioural data relative to
non-target values. Requires a data frame that (at least) has non-target value
data and participant response data.}
\usage{
plot_error_non_targets(
  data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  non_target_var = "non_target",
  set_size_var = NULL,
  condition_var = NULL,
  n_bins = 18,
  n_col = 2,
  return_data = FALSE
)
}
\arguments{
\item{data}{A data frame with columns containing: participant identifier
('id_var'); the participants' response per trial ('response_var'); the
non-target values ('non-target_var'); and, if applicable, the set size of each
response ('set_size_var'), and the condition of each response
('condition_var').}

\item{unit}{The unit of measurement in the data frame: "degrees"
(measurement is in degrees, from 0 to 360); "degrees_180 (measurement is in
degrees, but limited to 0 to 180); or "radians" (measurement is in radians,
from pi to 2 * pi, but could also be already in -pi to pi).}

\item{id_var}{The column name coding for participant id. If the data is from
a single participant (i.e., there is no id column) set to "NULL".}

\item{response_var}{The column name coding for the participants' responses.
The quoted variable name common to all columns (if applicable) storing
non-target values. The user should have one column coding for each
non-target's value in the data frame. If there is more than one non-target,
each column name should begin with a common term (e.g., the "non_target"
term is common to the non-target columns "non_target_1", "non_target_2"
etc.), which should then be passed to the function via the
\code{non_target_var} variable.}

\item{set_size_var}{The column name (if applicable) coding for the set
size of each response.}

\item{condition_var}{The column name (if applicable) coding for the
condition of each response.}

\item{n_bins}{An integer controlling the number of cells / bins used in the plot.}

\item{n_col}{An integer controlling the number of columns in the resulting
plot.}

\item{return_data}{A boolean (TRUE or FALSE) indicating whether the data for
the plot should be returned.}
}
\description{
Plot response error of behavioural target relative to non-target values
#'
Function to plot the response error in behavioural data relative to
non-target values. Requires a data frame that (at least) has non-target value
data and participant response data.
}
\examples{
data(bays2009_full)
plot_error_non_targets(bays2009_full,
                       unit = "radians",
                       id_var = "id",
                       response_var = "response",
                       non_target_var = "non_target",
                       set_size_var = "set_size",
                       condition_var = "duration")

}
