% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_model_fit}
\alias{plot_model_fit}
\title{Plot model fit against human error data (target errors)}
\usage{
plot_model_fit(
  human_data,
  model_data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  set_size_var = NULL,
  condition_var = NULL
)
}
\description{
Plot model fit against human error data (target errors)
}
