% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_parameters}
\alias{plot_parameters}
\title{Plot best-fitting parameters of model fit}
\usage{
plot_parameters(
  model_fit,
  id_var = "id",
  set_size_var = NULL,
  condition_var = NULL,
  n_col = 2,
  return_data = FALSE
)
}
\arguments{
\item{model_fit}{The model fit object containing the parameters to be
plotted.}

\item{id_var}{The column name coding for participant id. If the data is from
a single participant (i.e., there is no id column) set to NULL.}

\item{set_size_var}{The column name (if applicable) coding for the set
size of each response.}

\item{condition_var}{The column name (if applicable) coding for the
condition of each response.}

\item{n_col}{An integer controlling the number of columns in the resulting
plot.}

\item{return_data}{A boolean (TRUE or FALSE) indicating whether the data for
the plot should be returned.}
}
\description{
Function to plot the best-fitting parameters of either the 2-component or
3-component model. .
}
