% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_resultant_vector_length}
\alias{plot_resultant_vector_length}
\title{degrees, but limited to 0 to 180); or "radians" (measurement is in radians,
from pi to 2 * pi, but could also be already in -pi to pi).
a single participant (i.e., there is no id column) set to "NULL".
size of each response.
condition of each response.
the plot should be returned.}
\usage{
plot_resultant_vector_length(
  data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  set_size_var = NULL,
  condition_var = NULL,
  return_data = FALSE
)
}
\description{
degrees, but limited to 0 to 180); or "radians" (measurement is in radians,
from pi to 2 * pi, but could also be already in -pi to pi).
a single participant (i.e., there is no id column) set to "NULL".
size of each response.
condition of each response.
the plot should be returned.
}
