% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Transform radians into degrees}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://bayslab.com) published under GNU General Public License.
}
\usage{
radians_to_degrees(rad)
}
\arguments{
\item{rad}{Radian value to transform into degrees}
}
\description{
A function to transform radians into degrees
}
