% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sd2k}
\alias{sd2k}
\title{Translate from standard deviation to Von Mises K parameter
Returns the Von Mises concentration parameter K corresponding
to a standard deviation S of a wrapped normal distributions}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
sd2k(s)
}
\description{
Translate from standard deviation to Von Mises K parameter
Returns the Von Mises concentration parameter K corresponding
to a standard deviation S of a wrapped normal distributions
}
