% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulate_components}
\alias{simulate_components}
\title{simulate the mixture model. This function not to be called by the user}
\usage{
simulate_components(
  n_trials,
  kappa,
  p_u,
  p_n = NULL,
  set_size = 4,
  min_angle_distance = 20
)
}
\description{
simulate the mixture model. This function not to be called by the user
}
