% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulate_mixtur}
\alias{simulate_mixtur}
\title{simulate the mixture model. This is the function to be called by the user}
\usage{
simulate_mixtur(
  n_trials,
  K,
  p_t,
  p_n,
  p_u,
  set_size = 4,
  min_angle_distance = 20
)
}
\description{
simulate the mixture model. This is the function to be called by the user
}
