% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{slots_model_pdf_gd}
\alias{slots_model_pdf_gd}
\title{Calculate the likelihood function of the slots model fitting via
gradient descent (nelder-mead routine via optim function).}
\usage{
slots_model_pdf_gd(data, parms, min_parms, max_parms)
}
\description{
It is not expected that this function be called by the user.
}
