% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trapz}
\alias{trapz}
\title{trapz function from the caTools package by Jarek Tuszynski}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
trapz(x, y)
}
\description{
trapz function from the caTools package by Jarek Tuszynski
}
