% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vonmisespdf}
\alias{vonmisespdf}
\title{Probability density function of the Von Mises distribution.
Returns the probability density function for the Von Mises distribution with
mean MU and concentration K, evaluated at the values in X (given in
radians).}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
vonmisespdf(x, mu, k)
}
\description{
Probability density function of the Von Mises distribution.
Returns the probability density function for the Von Mises distribution with
mean MU and concentration K, evaluated at the values in X (given in
radians).
}
