% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap}
\alias{wrap}
\title{Map values onto circular space
This function maps the input radians onto the circular space -pi to pi.}
\usage{
wrap(data, bound = pi)
}
\arguments{
\item{data}{The data (in radians) to be mapped}
}
\description{
Map values onto circular space
This function maps the input radians onto the circular space -pi to pi.
}
