% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{A1inv}
\alias{A1inv}
\title{Inverse of A1 function.}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
A1inv(r)
}
\description{
Inverse of A1 function.
}
