% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_level}
\alias{fit_level}
\title{Fit model to a single level.}
\usage{
fit_level(
  data,
  components = 3,
  id_var = "id",
  response_var = "response",
  target_var = "target",
  non_target_var,
  set_size = 1,
  return_fit = FALSE,
  fit_method = "EM"
)
}
\description{
This wrapper function is called by the \code{fit_mixtur} function to fit the
models to a single level from the data frame. It is not expected that this
function be called by the user.
}
