% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{k2sd}
\alias{k2sd}
\title{Standard deviation of Von Mises K parameter
Returns the standard deviation of a wrapped normal distribution
corresponding to a Von Mises concentration parameter of K
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta}
\usage{
k2sd(k)
}
\description{
Standard deviation of Von Mises K parameter
Returns the standard deviation of a wrapped normal distribution
corresponding to a Von Mises concentration parameter of K
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta
}
