% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_precision}
\alias{plot_precision}
\title{Plot precision behavioural data}
\usage{
plot_precision(
  data,
  unit = "degrees",
  id_var = "id",
  response_var = "response",
  target_var = "target",
  set_size_var = "NULL",
  condition_var = "NULL",
  return_data = FALSE
)
}
\arguments{
\item{data}{A data frame with columns containing: participant identifier
('id_var'); the participants' response per trial ('response_var'); the
target value ('target_var'); and, if applicable, the set size of each
response ('set_size_var'), and the condition of each response
('condition_var').}

\item{unit}{The unit of measurement in the data frame: "degrees"
(measurement is in degrees, from 0 to 360); "degrees_180 (measurement is in
degrees, but limited to 0 to 180); "radians" (measurement is in radians,
from pi to 2 * pi); "wrapped_radians" (measurement is in radians, but
wrapped from -pi to pi)}

\item{id_var}{The column name coding for participant id.}

\item{response_var}{The column name coding for the participants' responses}

\item{target_var}{The column name coding for the target value}

\item{set_size_var}{The column name (if applicable) coding for the set
size of each response}

\item{condition_var}{The column name (if applicable) coding for the
condition of each response}

\item{return_data}{A boolean indicating whether the data for the plot should
be returned.}
}
\description{
This note is a TODO for later.
}
\examples{
library(tidyverse)
data(example_data)
plot_error(example_data, condition_var = "condition")

}
