% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{radians_to_degrees}
\alias{radians_to_degrees}
\title{Transform radians into degrees}
\usage{
radians_to_degrees(rad)
}
\arguments{
\item{deg}{Radian value to transform into degrees}
}
\description{
A function to transform radians into degrees
}
