% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{randomvonmises}
\alias{randomvonmises}
\title{Generates N random samples from a Von Mises distribution with mean mu and
concentration k}
\usage{
randomvonmises(n, mu, k)
}
\description{
Generates N random samples from a Von Mises distribution with mean mu and
concentration k
}
