% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sd2k}
\alias{sd2k}
\title{Translate from standard deviation to Von Mises K parameter
Returns the Von Mises concentration parameter K corresponding
to a standard deviation S of a wrapped normal distributions
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta
#' @export}
\usage{
sd2k(s)
}
\description{
Translate from standard deviation to Von Mises K parameter
Returns the Von Mises concentration parameter K corresponding
to a standard deviation S of a wrapped normal distributions
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta
#' @export
}
