% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sd2k}
\alias{sd2k}
\title{Translate from standard deviation to von Mises kappa parameter
Returns the von Mises concentration parameter kappa corresponding
to a standard deviation sd of a wrapped normal distribution}
\source{
The code has been adapted from Matlab code written by Paul Bays
(https://paulbays.com).
}
\usage{
sd2k(sd)
}
\arguments{
\item{sd}{A numeric value for standard deviation}
}
\description{
Translate from standard deviation to von Mises kappa parameter
Returns the von Mises concentration parameter kappa corresponding
to a standard deviation sd of a wrapped normal distribution
}
\examples{
kappa <- sd2k(0.361)
}
