% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulate_mixtur}
\alias{simulate_mixtur}
\title{Generate simulated data from mixture models}
\usage{
simulate_mixtur(n_trials, model, kappa, p_u, p_n, K, set_size)
}
\description{
Generate simulated data from mixture models
}
