% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulating.R
\name{simulate_slots}
\alias{simulate_slots}
\title{simulate the slots model. This is the function to be called by the user}
\usage{
simulate_slots(n_trials, model = "slots", K, kappa, set_sizes = c(1, 2, 4, 8))
}
\description{
simulate the slots model. This is the function to be called by the user
}
